I have connected all of the functions together. Users can now log in and get their code. I have also added time.sleep to let the pages load. If there is no delay, then the python will run too quickly, ultimately resulting it not finding elements.
Could not figure out how to indent. Users can possibly beautify their code themselves later, OR I could put it into an API or something to do it for them. Not a big concern right now though.

A big concern is that I get an Stale Element Error when trying to scrape a website(happens occasionally and i do not know why yet)